Draft Code of Conduct for Clear My Record
A standard Code of Conduct was downloaded as an initial drafting source. I added the paragraph entitled "Privacy" and will forward to Ellen for her review and approval.  We might consider a specific protocol for educating team members regarding obligations under the code or rules of professional conduct or responsibility.